fix: return type of amountFor method

+++ java/src/main/java/Customer.java
@@ -50,8 +50,8 @@ public class Customer {
         return result;
     }
 
-    private int amountFor(Rental each){
-        int thisAmount = 0;
+    private double amountFor(Rental each){
+        double thisAmount = 0;
         switch (each.getMovie().getPriceCode()) {
             case Movie.REGULAR:
                 thisAmount += 2;
